+++ include/game-models/AidKit/aid-kit.h
@@ -19,11 +19,11 @@ public:
 
 	void supply(std::shared_ptr<Player> player) override;
 
-	inline static const double SUPPLY_PORTION = 25.0;
+	inline static const int SUPPLY_PORTION = 25;
 
 private:
 	inline static const Vector2D SHAPE_COLLIDER_SIZE = Vector2D(10.0, 10.0);
-	inline static const double INITIAL_SUPPLY_CAPACITY = 200.0;
+	inline static const int INITIAL_SUPPLY_CAPACITY = 200;
 };
 
 } // namespace invasion::game_models
